Add api helper
